% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Monthly flow data from Estero Culebron DGA station}
\format{
A data frame with 480 rows and 5 variables:
\describe{
  \item{date}{dates}
  \item{year}{years}
  \item{month}{months}
  \item{day}{days}
  \item{4400001}{flows, in cubic meters per second}
}
}
\source{
\url{https://camels.cr2.cl/}
}
\usage{
data
}
\description{
A dataset containing dates and monthly flow values.
}
\keyword{datasets}
