% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_boxplot_package.R
\name{DR_boxplot}
\alias{DR_boxplot}
\title{BoxPlot from the result from the application of the Deductive Rational Method}
\usage{
DR_boxplot(dr)
}
\arguments{
\item{dr}{data frame from DR function}
}
\description{
BoxPlot from the result from the application of the Deductive Rational Method
}
