% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_plot_package.R
\name{DR_plot}
\alias{DR_plot}
\title{Plot the result from the application of the Deductive Rational Method}
\usage{
DR_plot(dr, facet = FALSE)
}
\arguments{
\item{dr}{data frame from DR function}

\item{facet}{logical, if TRUE, the plot will be faceted by year}
}
\description{
Plot the result from the application of the Deductive Rational Method
}
