% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_SVC_package.R
\name{DR_SVC}
\alias{DR_SVC}
\title{Seasonal Variation Curves considering results from the application of the Deductive Rational Method}
\usage{
DR_SVC(dr)
}
\arguments{
\item{dr}{data frame from DR function}
}
\description{
Seasonal Variation Curves considering results from the application of the Deductive Rational Method
}
