% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_timeseries_plot_package.R
\name{DR_timeseries_plot}
\alias{DR_timeseries_plot}
\title{Plot raw and infilled monthly time series}
\usage{
DR_timeseries_plot(raw_data, infilled_data)
}
\arguments{
\item{raw_data}{data frame with columns: date - year - month - day - station (Type date and numeric the rest). Assumes station data is in the 5th column.}

\item{infilled_data}{data frame from DR function (wide format: year, month1, month2, ...).}
}
\description{
Plot raw and infilled monthly time series
}
