% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_package.R
\name{RD}
\alias{RD}
\title{Apply the rational deductive method to a monthly series of flow or
precipitation data to fill in missing data.}
\usage{
RD(data)
}
\arguments{
\item{data}{data frame with columns: date - year - month - day - station (Type date and numeric the rest)}
}
\description{
Apply the rational deductive method to a monthly series of flow or
precipitation data to fill in missing data.
}
