% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_plot_package.R
\name{RD_plot}
\alias{RD_plot}
\title{Plot the result from the application of the rational deductive method}
\usage{
RD_plot(rd)
}
\arguments{
\item{rd}{data frame from RD function}
}
\description{
Plot the result from the application of the rational deductive method
}
