% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_stats_package.R
\name{RD_stats}
\alias{RD_stats}
\title{Statistics for the result from the application of the rational deductive method}
\usage{
RD_stats(rd)
}
\arguments{
\item{rd}{data frame from RD function}
}
\description{
Statistics for the result from the application of the rational deductive method
}
