% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_package.R
\name{DR}
\alias{DR}
\title{Apply the Deductive Rational Method to a monthly series of flow or
precipitation data to fill in missing data.}
\usage{
DR(data)
}
\arguments{
\item{data}{data frame with columns: date - year - month - day - station (Type date and numeric the rest)}
}
\description{
Apply the Deductive Rational Method to a monthly series of flow or
precipitation data to fill in missing data.
}
