% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_stats_package.R
\name{DR_stats}
\alias{DR_stats}
\title{Statistics for the result from the application of the Deductive Rational Method}
\usage{
DR_stats(dr)
}
\arguments{
\item{dr}{data frame from DR function}
}
\value{data_long_month: data frame with statistical summary of monthly value that result from the application of the Deductive Rational Method}
\description{
Statistics for the result from the application of the Deductive Rational Method
}

